requirements.txt
bcrypt==4.0.1
boto3==1.24.89
botocore==1.27.89
certifi==2022.9.24
cffi==1.15.1
cftime==1.6.2
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
dask==2022.2.1
fonttools==4.37.4
fsspec==2022.8.2
holidays==0.24
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
locket==1.0.0
matplotlib==3.6.1
netCDF4==1.6.1
numpy==1.21.6
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
partd==1.3.0
Pillow==9.2.0
prophet==1.1.1
pycparser==2.21
pylake==0.1.5
PyNaCl==1.5.0
pyparsing==3.0.9
pysftp==0.2.9
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
rasterio==1.3.6
requests==2.28.1
s3transfer==0.6.0
scipy==1.9.2
six==1.16.0
toolz==0.12.0
urllib3==1.26.12
xarray==0.20.2